refactor(iocounter): move iocounter package to internal
This is a utility package meant to implement a functionality for our own
code. It is not a public API. This moves it to internal to correctly
communicate that this isn't a formal part of the API contract.
