+++ internal/iocounter/writer.go
